+++ konva.js
@@ -3,7 +3,7 @@
  * Konva JavaScript Framework v0.11.1
  * http://konvajs.github.io/
  * Licensed under the MIT or GPL Version 2 licenses.
- * Date: Sat Jan 16 2016
+ * Date: Tue Jan 26 2016
  *
  * Original work Copyright (C) 2011 - 2013 by Eric Rowell (KineticJS)
  * Modified work Copyright (C) 2014 - 2015 by Anton Lavrenov (Konva)
@@ -4022,7 +4022,7 @@
             var events = this.eventListeners[eventType],
                 i;
 
-            evt = Konva.Util.cloneObject(evt || {});
+            evt = evt || {};
             evt.currentTarget = this;
             evt.type = eventType;